refactor: remove navigation array and render component per section
