+++ app/services/cms/models/StrapiContentComponent.ts
@@ -11,7 +11,7 @@ import { StrapiInlineNoticeSchema } from "./StrapiInlineNotice";
 import { StrapiLinkListBoxSchema } from "./StrapiLinkListBox";
 import { StrapiListSchema } from "./StrapiList";
 import { StrapiParagraphSchema } from "./StrapiParagraph";
-import { StrapiSummaryOverviewSchema } from "./StrapiSummaryOverview";
+import { StrapiSummaryOverviewSectionSchema } from "./StrapiSummaryOverviewSection";
 import { StrapiUserFeedbackSchema } from "./StrapiUserFeedback";
 
 export const StrapiContentComponentSchema = z.union([
@@ -28,7 +28,7 @@ export const StrapiContentComponentSchema = z.union([
   StrapiInlineNoticeSchema,
   StrapiDetailsSchema,
   StrapiUserFeedbackSchema,
-  StrapiSummaryOverviewSchema,
+  StrapiSummaryOverviewSectionSchema,
 ]);
 
 export type StrapiContentComponent = z.infer<
